requirements.txt
fastapi
uvicorn
azure-messaging-webpubsubservice
python-dotenv